feat: Adds smoke tests that attempt to fetch 'GET /' and 'GET /404'
